update source code 20180208

+++ kernel-3.18/drivers/misc/mediatek/teei/V1.0/tz_driver/include/global_function.h
@@ -4,7 +4,10 @@ extern int forward_call_flag;
 extern int irq_call_flag;
 extern int fp_call_flag;
 extern int keymaster_call_flag;
-extern int teei_vfs_flag;
+//zhangheting@wind-mobi.com add teei patch 20170523 start
+//extern int teei_vfs_flag;
+extern unsigned long teei_vfs_flag;
+//zhangheting@wind-mobi.com add teei patch 20170523 end
 extern struct completion global_down_lock;
 extern unsigned long teei_config_flag;
