+++ kernel-3.18/drivers/misc/mediatek/teei/V1.0/tz_driver/switch_queue.c
@@ -56,7 +56,10 @@ extern struct kthread_worker ut_fastcall_worker;
 extern int forward_call_flag;
 extern int fp_call_flag;
 extern int keymaster_call_flag;
-extern int teei_vfs_flag;
+//zhangheting@wind-mobi.com add teei patch 20170523 start
+//extern int teei_vfs_flag;
+extern unsigned long teei_vfs_flag;
+//zhangheting@wind-mobi.com add teei patch 20170523 end
 extern int irq_call_flag;
 
 extern void nt_sched_t_call(void);
